backend/requirements.txt
Flask
fastapi
uvicorn
openai
pinecone-client
numpy
pandas
scikit-learn
python-dotenv
requests